introduce fragment_features.py and features_utils.py
